add 'OTP without password' support to RADIUS PAP protocol
If 'OTP without password' is enabled for PAP in radius.conf, an OTP token code is always required for authentication via PAP, even if the account has no OTP token activated.
This is useful to prevent authentication with the password as the OTP token code if the account has no token, but differs from the behaviour of CHAP and MSCHAPV2.

- add 'OTP without password' support to PAP implementation
- allow individual configuration 'OTP without password' for CHAP, MSCHAPV2 and PAP for backward compatibility
- update documentation