fix realm retrieval by name vs id
RealmProvider can retrieve realms by ID (usually a UUID) or by name (the realm name specified e.g. when logging in with the username). The realm information from the RADIUS packages is always the name and therefore must be retrieved with RealmProvider#getRealmByName() instead of RealmProvider#getRealm(). (I'm not sure how this worked properly before and what change introduced the problem.)
The test was also enhanced and adapted accordingly.